Move dispatch methods into GenericApp to avoid circular dependencies.
Clean up various argument orders.

Simplify use of methods that aren't dispatched.
