+++ morepath/path.py
@@ -15,6 +15,8 @@ except ImportError:
     from urllib.parse import urlencode, quote
 
 
+from .generic import GenericApp
+from .reify import reify
 from .cachingreg import RegRegistry
 from .traject import Path as TrajectPath, TrajectRegistry
 from .converter import ParameterFactory, ConverterRegistry, IDENTITY_CONVERTER
@@ -50,6 +52,14 @@ class PathRegistry(TrajectRegistry):
         self.converter_registry = converter_registry
         self.mounted = {}
         self.named_mounted = {}
+
+    @reify
+    def lookup(self):
+        """Get the :class:`reg.Lookup` for this application.
+
+        :return: a :class:`reg.Lookup` instance.
+        """
+        return self.reg_registry.caching_lookup
 
     def register_path(self, model, path,
                       variables, converters, required, get_converters,
@@ -141,9 +151,7 @@ class PathRegistry(TrajectRegistry):
         :param func: function that gets a model instance argument and
           returns a variables dict.
         """
-        # XXX
-        from .app import App
-        self.reg_registry.register_function(App._path_variables,
+        self.reg_registry.register_function(GenericApp._path_variables,
                                             func,
                                             obj=model)
 
@@ -161,16 +169,15 @@ class PathRegistry(TrajectRegistry):
         converters = converters or {}
         get_path = Path(path, factory_args, converters, absorb)
 
-        # XXX
-        from .app import App
-        self.reg_registry.register_function(App._class_path, get_path,
+        self.reg_registry.register_function(GenericApp._class_path, get_path,
                                             model=model)
 
         def default_path_variables(obj):
             return {name: getattr(obj, name) for name in factory_args}
-        self.reg_registry.register_function(App._default_path_variables,
-                                            default_path_variables,
-                                            obj=model)
+        self.reg_registry.register_function(
+            GenericApp._default_path_variables,
+            default_path_variables,
+            obj=model)
 
     def register_defer_links(self, model, app_factory):
         """Register factory for app to defer links to.
@@ -182,9 +189,8 @@ class PathRegistry(TrajectRegistry):
           object as arguments and should return another app instance that
           does the link generation.
         """
-        from .app import App
         self.reg_registry.register_function(
-            App._deferred_link_app, app_factory,
+            GenericApp._deferred_link_app, app_factory,
             obj=model)
 
     def register_defer_class_links(self, model, get_variables, app_factory):
@@ -199,10 +205,8 @@ class PathRegistry(TrajectRegistry):
           app instance that does the link generation.
         """
         self.register_path_variables(model, get_variables)
-        # XXX
-        from .app import App
         self.reg_registry.register_function(
-            App._deferred_class_link_app, app_factory,
+            GenericApp._deferred_class_link_app, app_factory,
             model=model)
 
 
@@ -302,7 +306,7 @@ class Path(object):
 
         :param model: model class. Not actually used in the
           implementation but used for dispatch in
-          :meth:`App._class_path`.
+          :meth:`GenericApp._class_path`.
         :param variables: dict with the variables used in the path. each
           argument to the factory function should be represented.
         :return: :class:`PathInfo` instance representing the path.
