+++ morepath/tests/test_security.py
@@ -2,7 +2,6 @@
 import pytest
 import morepath
 from morepath.request import Response
-#from morepath import generic
 from morepath.authentication import Identity, NO_IDENTITY
 from .fixtures import identity_policy
 import base64
