+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist = py27,pypy,py33,py34,py35,pep8,coverage,docs
 [testenv]
-deps=git+https://github.com/morepath/reg.git#egg=reg
+deps=git+https://github.com/morepath/reg.git@dispatchmethods#egg=reg
      -e{toxinidir}[test]
      -e{toxinidir}/fixture_packages/base
      -e{toxinidir}/fixture_packages/sub
